creating a subscribe form common component to hold better the translations
