+++ frontend/src/translations/en.json
@@ -86,6 +86,7 @@
   "account.profile.welcome.description": "This is the first steps to start to work with Gitpay",
   "account.profile.tasks.caption": "Tasks",
   "account.profile.tasks.headline": "Tasks",
+  "account.profile.tasks.setup": "Tasks",
   "account.profile.tasks.description": "Check the tasks available for you",
   "account.profile.tasks.link.tasks": "See tasks",
   "account.profile.tasks.payment.caption": "Payment",
@@ -128,7 +129,7 @@
   "task.payment.headline.bounty.add": "Add a bounty for this task",
   "task.payment.form.message.subheading": "Create a bounty for this task and who you assign will receive the payment for this bounty",
   "task.payment.input.amount.value": "Price",
-  "task.payment.input.amount": "Price insert a value for this task",
+  "task.payment.input.amount": "Add a bounty for this task",
   "task.payment.creditcard.action": "Pay $ {amount} with Credit Card",
   "task.payment.paypal.action": "Pay $ {amount} with PayPal",
   "task.status.label.open": "Open",
@@ -148,7 +149,7 @@
   "task.payment.transfers.concluded": "All transfers was succeeded to the destination account",
   "task.payment.transfer.done": "All your transfer was concluded with your credit card and the transaction id is:",
   "task.payment.pay.button.credit": "Pay $ {value}",
-  "task.payment.pay.button.paypal": "Pay with PayPal (id: {transfer}",
+  "task.payment.pay.button.paypal": "Pay with PayPal (id: {transfer})",
   "task.payment.done.to": "You made a payment to $ {user}",
   "task.payment.start.payTo": "Pay $ {value}",
   "task.payment.types.notype": "No payment for this payment type",
@@ -251,7 +252,7 @@
   "welcome.main.item.one.primary": "Work in projects using the best development tools",
   "welcome.main.item.one.secondary": "We use all the Git tools and version control to manage deliveries for our clients",
   "welcome.main.item.two.primary": "Colaboration with companies by tasks on demand",
-  "welcome.main.item.two.secondary": "Work in different projects, you can colaborate and learn with many projects and stacks",
+  "welcome.main.item.two.secondary": "Work in different projects, colaborate and learn with many projects and stacks",
   "welcome.main.item.three.primary": "Receive bounties by colaboration",
   "welcome.main.item.three.secondary": "Receive bounties for the task you concluded with direct payment when your code is merged on the codebase",
   "welcome.companies.item.one.primary": "Manage the tasks of your projects",
@@ -314,5 +315,17 @@
   "consts.account.statuses.finish": "Finished",
   "consts.account.statuses.undefined": "Undefined",
   "consts.account.statuses.undefined.two": "Undefined",
-  "consts.account.statuses.undefined.three": "Undefined"
+  "consts.account.statuses.undefined.three": "Undefined",
+  "form.email.leave.label": "Leave your email",
+  "form.email.signup.label": "Signup!",
+  "form.email.subscribing.label": "Subscribing...",
+  "form.email.submit.success": "You now registered and soon you will receive new challenges",
+  "form.email.submit.error": "We could not register your e-mail, maybe you're already registered?",
+  "info.stats.number.tasks": "tasks",
+  "info.stats.number.bounty": "paid for bounties",
+  "info.stats.number.users": "users",
+  "welcome.how.consulting.main.title": "Consultancy",
+  "welcome.how.consulting.who.subscribers": "Gitpay is for you?",
+  "welcome.how.workflow.main.title": "Workflow",
+  "welcome.how.workflow.contrib.title": "For contributors"
 }
